Rename directory devnet -> devnet_utils
